Added unit test for both repositories.
