+++ SampleStack.Repository.Tests/Repositories/TestRecord.cs
@@ -1,0 +1,10 @@
+﻿using SampleStack.Repository.Models;
+
+namespace SampleStack.Repository.Repositories.Tests
+{
+    public record TestRecord : IModelBase
+    {
+        public int Id { get; set; }
+        public string? Name { get; set; }
+    }
+}
